+++ Sources/Parsing/BlockScalarChomping.swift
@@ -1,0 +1,10 @@
+extension PureYAML.Parsing {
+    enum BlockScalarChomping: String, Equatable, CustomStringConvertible {
+        case clip
+        case strip
+
+        var description: String {
+            rawValue
+        }
+    }
+}